[ft-#162339289] build out end-point for users to get a specific record and delete a specific record
