+++ test/testGetIncidentRecord.js
@@ -33,6 +33,9 @@ describe("the root endpoint", () => {
   });
 });
 
+/**
+ * get all red-flags
+ */
 describe("GET all red-flag records endpoint", () => {
 
   it("should return status 200 if red-flag records exist in the database", (done) => {
@@ -53,16 +56,23 @@ describe("GET all red-flag records endpoint", () => {
 
 });
 
+
+
+/**
+ * Get specific red-flag
+ */
 describe("GET a specific red-flag record endpoint", () => {
 
   it("should return status 200 if specified red-flag record exist in the database", (done) => {
     request
       .get(`${red_flags}/${validURI}`)
+      .set("authorization", validToken)
       .end((err, res) => {
         expect(res.status).to.eql(200);
         expect(res.body.status).to.eql(200);
         expect(res.body.status).to.be.a("number");
         (res.body.data[0]).should.be.an("object");
+        (res.body.data[0].id).should.be.an("number");
         should.not.exist(err);
         should.exist(res.body);
         if (err) { return done(err); }
@@ -73,6 +83,7 @@ describe("GET a specific red-flag record endpoint", () => {
   it("should return status 404 if specified red-flag record doesn't exist in the database", (done) => {
     request
       .get(`${red_flags}/${invalidURI}`)
+      .set("authorization", validToken)
       .end((err, res) => {
         expect(res.status).to.eql(404);
         expect(res.body.error).to.eql("record not found");
